Moves biome detection into MapRegion.  Updates test models and API docs.
